Бот теперь общается с каждым пользователем индивидуально, включая показ списка покупок, исправил файловый набор и набор директорий, убрал лишние хендлеры (из трека)
