+++ shopping_manager_bot/bot/__init__.py
@@ -1,0 +1,1 @@
+
