+++ shopping_manager_bot/db/__init__.py
@@ -1,0 +1,1 @@
+
