+++ shopping_manager_bot/db/db.py
@@ -1,12 +1,9 @@
-import os, sys
-sys.path.insert(1, os.path.join(sys.path[0], '..'))
-from settings.settings_file import URL
 from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker, scoped_session
+from settings_box.settings import URL
 
 engine = create_engine(URL)
 db_session = scoped_session(sessionmaker(bind=engine))
-
 Base = declarative_base()
 Base.query = db_session.query_property()
