--- shopping_manager_bot/settings/settings_file.py
@@ -1,3 +1,0 @@
-API_KEY = '5941807026:AAHmB8Afi3zf7inQd00zUwtLD003MEL7Ozg'
-URL = "postgresql://cjpzyqur:oSZkJ8wmL8FFUG5Pn1gy1bywlsqAoslH@mel.db.elephantsql.com/cjpzyqur"
-USER_EMOJI = [":shit:", ":smile:", ":sunglasses:", ":eyes:"]
